Refactor windowInfo for consistency & readability
Introduced several new varaibles for active scene and windowed size,
so this can now be accessed in pixels or in tiles. This can also be
automaticallyset so that the pixel size is the correct multiple of
the tile size.

Introduced setPosition function which can set the position on the
window as an offset given in pixels from either an alignment on the
window (top, bottom, centre, left, right), or from the x & y
position of an object on the screen

Reduced line length for some of the more egrigious setposition lines
by using temporary x & y position variables for readability