Move BattlingCharacter stats into .json
BattlingCharacter stats now stored in a .json format and can be modified in data. This has been implemented for the player class but other BattlingClasses are not yet written. In the future this will allow implementation of new enemies purely in data, as I intend to make behaviour methods accessible in data and to also define animations in data