Decouple behaviour triggers from key inputs
BattlingCharacter behaviour triggers are now independent from the key inputs, so they can be triggered either by key inputs for player or by AI for enemies. In addition CardsInHand will now only render cards when the identifier is the player. In future should instead change this behaviour so it renders the enemy cards to UI in a different way to the player.