Add pathfinding code to enemy
Enemies can now pathfind between points using A*. This version creates a test enemy whose only action is to find a path between itself and the player. It then passes this to the tile map which draws it as a card action for testing. Currently there is some kind of bug where it fails if you are too far to the right or to the bottom. There is also a bug where the screen scrolls incorrectly. Hopefully these will both be fixed next commit.