Bug fix - Memory write access violation in debug mode
Running in debug mode would cause a write access violation. Eventually I tracked this down to the type pointer in player pointing to garbage after leaving the game constructor. This was because BattlingCharacterTypes was being passed by value and so destroyed after leaving the scope. I'm assuming the optimisations in release mode were smart enough to detect that this was not modified inside the game constructor therefore passed it by value regardless