Tilemap now uses a vertex array alongside a single tileset texture for drawing the background.
This should result in (hopefully significantly) improved performance. This was done as the previous method (updating a texture based on multiple images) was against SFML best practices as updating a texture is a relatively heavy operation. It also has the advantage of reducing the number of .png files by allowing the whole tileset to be contained in a single image, which is loaded into memory as a single texture