Groundwork for different animation state classes
Adds an abstract base class State Animation. The virtual function updateAnimation will take various parameters depending on the animation type, and will update those parameters according to the keyframe data stored in keyframes. Keyframes themselves don't care about how the data in them is interpreted, this is determined by the derived state animation class that they are used in. Aside from the actual frame number (although this is not actually stored as a frame but as a time in seconds after the animation is activated), keyframes can store an identifier string, a 2d float vector, and a RGBA colour.

Also added an animation manager class. Update animations does nothing right now but will eventually iterate through a vector of Playing Animations, a struct which stores both a pointer to the animation itself as well as the data about the current active time. The playing animation update function currently does nothing but will soon be overloaded with the various data types needed for the 3 different planned animation classes derived from the state animation class. Namely, Sprite animation (which will handle swapping out textures on the sprite), squash animation (handling squish and stretch of the sprite via transforms), and colour animation (colouring and transparency.