Eliminate warnings and clarify game draw function slightly
Wanrings have been largely eliminated by switching most for loop ints to unsigned ints, and by statically casting ints to floats. I have also clarified the draw function in the game class by adjusting some of the parameters to make it more clear what they are. In the future I should bring much of this code within the battling character class, as it makes sense for the characters to 'own' there own health an card point bars and be responsible for drawing them