Entities can now hold other entities as components.
Currently this is limited to a positional entity holding its sprite, and an animation manager. Eventually I would like to factor out all functionality into components. Entities positions can be defined globally or locally relative to the parent entity.

This code compiles but currently hits an unhandled exception on start. The components are currently also stored as raw pointers and will leak when the entity is destroyed, which I need to fix up.